requirements.txt
pandas==2.2.2
numpy==2.1.1
ccxt==4.4.48
pyyaml==6.0.2
pydantic==2.9.2
flask==3.0.3
loguru==0.7.2
python-dotenv==1.0.1
joblib==1.4.2
matplotlib==3.9.2
scipy==1.13.1
pytest==8.3.2
pytest-cov==5.0.0
mypy==1.11.2
flake8==7.1.1
requests==2.32.3
ctrader-open-api==0.0.27
